refactor(api-server): extract scheduler vote-lifecycle jobs to submodule
Churn-reduction follow-up to #2610 (retention/prune extraction). The
`scheduler/mod.rs` hotspot still carried the full vote expiry/reminder
cluster inline. Move the eight vote-lifecycle methods
(activate_scheduled_votes, close_expired_votes, send_vote_reminders and
their private target-resolution helpers) into a dedicated
`scheduler/votes.rs`, attached to `Scheduler` via a `impl` block, mirroring
the `retention` submodule pattern.

Pure move — no behaviour change. The three jobs invoked from
`run_scheduled_tasks` are exposed `pub(super)`; the helpers stay private to
the new module. mod.rs shrinks by ~385 lines (2525 -> 2134).
